+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Change log
 
+## v0.1.10 (2019-11-09)
+
+ - Update `create repo` command with support for `--defaultbranch "develop"`
+ - Create `update repo` command analogy with `create repo`
+
 ## v0.1.9 (2019-11-08)
 
  - Add go style checking (golint) and vet to pipeline